+++ samples/sampler/stories/default/WizardPanels.js
@@ -27,6 +27,8 @@ storiesOf('Sandstone', module)
 				onTransition={action('onTransition')}
 				onWillTransition={action('onWillTransition')}
 				prevButtonText={text('prevButtonText', WizardPanels, '')}
+				subtitle={text('subtitle', WizardPanels, '')}
+				title={text('title', WizardPanels, '')}
 			>
 				<WizardPanels.Panel footer="Footer in View 1" subtitle="A subtitle for View 1" title="WizardPanel View 1">
 					<Scroller>
@@ -37,7 +39,7 @@ storiesOf('Sandstone', module)
 						<Button>Cancel</Button>
 					</buttons>
 				</WizardPanels.Panel>
-				<WizardPanels.Panel subtitle="A subtitle for View 2" title="WizardPanel View 2">
+				<WizardPanels.Panel subtitle="A subtitle for View 2 that is really, really way too long for its own good.  In fact, it's so long that it probably goes to multiple lines, unless your screen is so large that it somehow fits.  That seems unlikely, though, unless you're in the year 2030 or something." title="WizardPanel View 2">
 					<BodyText>Several buttons!</BodyText>
 					<Button icon="list">Button A</Button>
 					<Button icon="gear">Button B</Button>